base_test_platform_agent renamed from base_test_platform_agent_with_rsn (mainly to avoid confusion as the tests are against the CI RSN OMS simulator, and not against the actual RSN OMS endpoint)
